Code improvements

+++ HTTPShortcuts/app/src/main/java/ch/rmy/android/http_shortcuts/realm/RealmFactory.kt
@@ -10,7 +10,7 @@ internal object RealmFactory {
 
     private val configuration: RealmConfiguration by lazy {
         RealmConfiguration.Builder()
-                .schemaVersion(DatabaseMigration.VERSION.toLong())
+                .schemaVersion(DatabaseMigration.VERSION)
                 .migration(DatabaseMigration())
                 .build()
     }
